pyproject.toml
[tool.poetry]
name = "Soul-Engine"
version = "0.0.1"
description = "Soul Engine is a real-time visualization and simulation engine built on the back of CUDA, OpenCL, and Vulkan."

license = "GPLv3"

authors = [
    "Synodic Software"
]

readme = "README.md"

repository = "https://github.com/Synodic-Software/Soul-Engine"
homepage = "https://github.com/Synodic-Software/Soul-Engine"

packages = [
    { include = "Scripts.py", from = "Tools/Python"}
]


[tool.poetry.dependencies]
python  = "3.7.*"


[tool.poetry.dev-dependencies]
conan   = "^1.18.*"


[tool.poetry.scripts]
setup   = "Scripts:Setup"
update  = "Scripts:Update"


[build-system]
requires        = ["poetry>=0.12"]
build-backend   = "poetry.masonry.api"